Do not export typescript files
Because the npm package currently outputs the typescript files along with the distro js/.map.js files, webpack loaders require loading in the typescript files directly. This increases webpack build times for loaders like ts-loader because they cannot safetly ignore the node_modules directory. This makes the necessary changes to not export the typescript files along with the distro files.

- Only export files in the dist directory.
- Place ES6 UMD bundle in the same directory as the dist/index.js (Require statement in the bundle file requires it be at the same level as its source which was dist/index.js).
- Fix issue with tsconfig.json regarding __decorate helper method missing.